requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.104.2
python-dotenv==1.1.1
aiosqlite==0.20.0
